# api/openapi.yaml
openapi: 3.1.0
info:
  title: Prompt Engine API
  description: 設計書§13 API設計（REST）
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- ciap-bearer: []
paths:
...
  /api/v1/prompts/{namespace}/{name}/versions:
    post:
      tags:
      - version-controller
      operationId: createVersion
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVersionRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/KeySemVerResponseDto"
  /api/v1/prompts/{namespace}/{name}/versions/{version}/publish:
    post:
      tags:
      - version-controller
      operationId: publish
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/KeySemVerResponseDto"
  /api/v1/prompts/{namespace}/{name}/versions/{version}/deprecate:
    post:
      tags:
      - version-controller
      operationId: deprecate
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeprecateRequestDto"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/KeySemVerResponseDto"
  /api/v1/prompts/{namespace}/{name}/rollback:
    post:
      tags:
      - version-controller
      operationId: rollback
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RollbackRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RollbackResponseDto"
...
  /api/v1/prompts/{namespace}/{name}:
    get:
      tags:
      - prompt-controller
      operationId: get
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PromptDetailDto"
    delete:
      tags:
      - prompt-controller
      operationId: archive
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: string
      - name: force
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
    patch:
      tags:
      - prompt-controller
      operationId: updateMetadata
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePromptMetadataRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/KeyOnlyResponseDto"
  /api/v1/prompts/{namespace}/{name}/versions/{version}:
    get:
      tags:
      - version-controller
      operationId: getVersion
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PromptVersionDto"
  /api/v1/prompts/{namespace}/{name}/diff:
    get:
      tags:
      - version-controller
      operationId: diff
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: from
        in: query
        required: true
        schema:
          type: string
      - name: to
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DiffResponseDto"
...
components:
  schemas:
    ContextRequirementDto:
      type: object
      properties:
        scope:
          type: string
        required:
          type: array
          items:
            type: string
        optional:
          type: array
          items:
            type: string
      required:
      - optional
      - required
      - scope
...
    OutputDeclarationDto:
      type: object
      properties:
        format:
          type: string
        schemaRef:
          type: string
      required:
      - format
    ValidationSettingsDto:
      type: object
      properties:
        maxLength:
          type: integer
          format: int32
        maxTokens:
          type: integer
          format: int32
        policies:
          type: array
          items:
            type: string
        placeholders:
          type: string
      required:
      - placeholders
      - policies
    VariableDefinitionDto:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        source:
          type: string
        required:
          type: boolean
        default:
          type: object
        constraints:
          type: array
          items:
            type: string
        sensitive:
          type: boolean
      required:
      - constraints
      - name
      - required
      - sensitive
      - source
      - type
    KeySemVerResponseDto:
      type: object
      properties:
        key:
          type: string
        semVer:
          type: string
      required:
      - key
      - semVer
    CreateVersionRequestDto:
      type: object
      properties:
        semVer:
          type: string
          minLength: 1
        source:
          type: string
          minLength: 1
        variables:
          type: array
          items:
            $ref: "#/components/schemas/VariableDefinitionDto"
        contextRequirements:
          type: array
          items:
            $ref: "#/components/schemas/ContextRequirementDto"
        validation:
          $ref: "#/components/schemas/ValidationSettingsDto"
        output:
          $ref: "#/components/schemas/OutputDeclarationDto"
      required:
      - contextRequirements
      - semVer
      - source
      - variables
    DeprecateRequestDto:
      type: object
      properties:
        recommendedReplacement:
          type: string
    RollbackRequestDto:
      type: object
      properties:
        targetVersion:
          type: string
          minLength: 1
      required:
      - targetVersion
    RollbackResponseDto:
      type: object
      properties:
        key:
          type: string
        targetSemVer:
          type: string
      required:
      - key
      - targetSemVer
...
    UpdatePromptMetadataRequestDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
      - name
      - tags
    KeyOnlyResponseDto:
      type: object
      properties:
        key:
          type: string
      required:
      - key
...
    ExtendsRefDto:
      type: object
      properties:
        key:
          type: string
        range:
          type: string
      required:
      - key
    PromptDetailDto:
      type: object
      properties:
        metadata:
          $ref: "#/components/schemas/PromptMetadataDto"
        versions:
          type: array
          items:
            $ref: "#/components/schemas/PromptVersionDto"
      required:
      - versions
    PromptMetadataDto:
      type: object
      properties:
        key:
          type: string
        name:
          type: string
        category:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
      - key
      - name
      - tags
    PromptVersionDto:
      type: object
      properties:
        semVer:
          type: string
        state:
          type: string
        source:
          type: string
        contentHash:
          type: string
        variables:
          type: array
          items:
            $ref: "#/components/schemas/VariableDefinitionDto"
        contextRequirements:
          type: array
          items:
            $ref: "#/components/schemas/ContextRequirementDto"
        extends:
          $ref: "#/components/schemas/ExtendsRefDto"
        validation:
          $ref: "#/components/schemas/ValidationSettingsDto"
        output:
          $ref: "#/components/schemas/OutputDeclarationDto"
      required:
      - contentHash
      - contextRequirements
      - semVer
      - source
      - state
      - validation
      - variables
    DiffResponseDto:
      type: object
      properties:
        key:
          type: string
        from:
          type: string
        to:
          type: string
        contentChanged:
          type: boolean
        fromContentHash:
          type: string
        toContentHash:
          type: string
        variablesChanged:
          type: boolean
        contextRequirementsChanged:
          type: boolean
        extendsChanged:
          type: boolean
        validationChanged:
          type: boolean
        outputChanged:
          type: boolean
      required:
      - contentChanged
      - contextRequirementsChanged
      - extendsChanged
      - from
      - fromContentHash
      - key
      - outputChanged
      - to
      - toContentHash
      - validationChanged
      - variablesChanged
...
  securitySchemes:
    ciap-bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT